~Version: 1.1.0
Now able to log in, onto the website, with the credentials, by the database provided. Still gotta make the website secure, in a way, that when a user is created, the pass needs to be hashed, and stored in the database, instead of hardcoded passwords, which might or might not get implemented in time.