~Version: 1.3.0
Tried fixing the login section, instead of logging in with plaintext, the password should be compared to a hash in the DB, as far as i've come, i get all the correct information i need, i just dont compare the hashes correctly, takes too long to finish, as i can't get it. Hopefully a feature tomorrow.. Otherwise, added support and alert messages to display for client users.